Add programmatic API
