+++ packages/core/src/constants.ts
@@ -1,2 +1,7 @@
 export const frontmatterFileExtensions = ['.yml', '.md', '.markdown', '.mdx', '.mdoc'];
 export const moduleFileExtensions = ['.js', '.mjs', '.cjs', '.ts', '.mts', '.cts'];
+export const errorOpts = {
+	prefix: '- ',
+	prefixSeparator: '',
+	issueSeparator: '\n- ',
+};
