+++ packages/core/src/index.ts
@@ -1,10 +1,4 @@
-import type { LunariaUserRendererConfig } from './types.js';
-
 export { nothing } from 'lit';
 export { html, svg, unsafeStatic } from 'lit/static-html.js';
-
+export { defineRendererConfig, lunaria } from './api.js';
 export type * from './types.js';
-
-export function defineRendererConfig(opts: LunariaUserRendererConfig) {
-	return opts;
-}
